GEOT-2681 fixed - integer semantic action was changed to handle long values
git-svn-id: http://svn.osgeo.org/geotools/trunk@33835 e5c1c795-43da-0310-a71f-fac65c449510

+++ modules/library/cql/src/main/java/org/geotools/filter/text/commons/AbstractFilterBuilder.java
@@ -366,7 +366,7 @@ public abstract class AbstractFilterBuilder {
 
     public Literal buildLiteralInteger(final String image) {
         
-        return filterFactory.literal(Integer.parseInt(image));
+        return filterFactory.literal(Long.parseLong(image));
     }
     public Literal buildLiteralDouble(final String tokenImage) {
